+++ lib/json_render.dart
@@ -3,3 +3,4 @@ library json_render;
 export 'src/json_render_base.dart';
 export 'src/json_render_media.dart';
 export 'src/json_render_css.dart';
+export 'src/json_render_charts.dart';